configuting default testnet
