+++ commons/ava_default_testnet/staker_tls_configs.go
@@ -1,58 +1,5 @@
 package ava_default_testnet
 
-/*
-	For the default testnet, there are five bootstrappers. They have hardcoded bootstrapperIDs that correspond to their TLS certs.
-	This must be hardcoded because Gecko requires specifying the bootstrapperID
-	along with the bootstrapperIP when connecting to bootstrappers in TLS mode.
-
-	The hardcoded IDs in this file are the known Gecko ID for a node using the private key and cert in this file.
-	They are also hardcoded in the gecko source code as the IDs for the initial stakers of the default testnet.
-*/
-
-type StakerIdentity struct {
-	NodeID string
-	PrivateKey string
-	TlsCert string
-}
-
-var DefaultStakers = []StakerIdentity{
-	Staker1,
-	Staker2,
-	Staker3,
-	Staker4,
-	Staker5,
-}
-
-var Staker1 = StakerIdentity{
-	STAKER_1_NODE_ID,
-	STAKER_1_PRIVATE_KEY,
-	STAKER_1_CERT,
-}
-
-var Staker2 = StakerIdentity{
-	STAKER_2_NODE_ID,
-	STAKER_2_PRIVATE_KEY,
-	STAKER_2_CERT,
-}
-
-var Staker3 = StakerIdentity{
-	STAKER_3_NODE_ID,
-	STAKER_3_PRIVATE_KEY,
-	STAKER_3_CERT,
-}
-
-var Staker4 = StakerIdentity{
-	STAKER_4_NODE_ID,
-	STAKER_4_PRIVATE_KEY,
-	STAKER_4_CERT,
-}
-
-var Staker5 = StakerIdentity{
-	STAKER_5_NODE_ID,
-	STAKER_5_PRIVATE_KEY,
-	STAKER_5_CERT,
-}
-
 const (
 	STAKER_1_NODE_ID = "7Xhw2mDxuDS44j42TCB6U5579esbSt3Lg"
 	STAKER_2_NODE_ID = "MFrZFVCXPv5iCn6M9K6XduxGTYp891xXZ"
@@ -476,3 +423,4 @@ zg/ibrLgatzA9SKgoWXm2bLt6bbXefmOgnZXyw8Qko70Xxtx5eBR1BDAQjDis81n
 Lg96sJ3LOn7SXHfxJ3BtXshTJAoBFx6EpmulgNoPWIkJtd7XWYP6Yy22D+kK7OhH
 Rq3CiYMtDmZoub/kVBL0MVdSm7hn1TSVTHjFoW6cwQ37iKHjkZVRwX1Kzt0B
 -----END RSA PRIVATE KEY-----`
+
